requirements.txt
alembic==0.8.3
autoenv==1.0.0
blinker==1.3
boto==2.43.0
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-Session==0.2.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.11
funcy==1.7.2
gunicorn==19.3.0
htmlmin==0.1.10
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
passlib==1.6.5
psycopg2==2.6.1
pydash==3.4.7
pydub==0.16.6
python-dateutil==2.6.0
python-editor==1.0.1
readline==6.2.4.1
requests==2.9.1
simplejson==3.8.1
six==1.10.0
SQLAlchemy==1.0.11
Werkzeug==0.11
WTForms==2.0.2
